Update pro 6.py

+++ pro 6.py
@@ -57,23 +57,23 @@ def display_menu():
              except FileNotFoundError:
                   print("No journal file found.")
 
-      elif choice == 4:
-               confirm = input(
-                   "Are you sure you want to delete all entries? (yes/no): ")
+       elif choice == 4:
+            confirm = input(
+                "Are you sure you want to delete all entries? (yes/no): ")
                
 
-               if confirm.lower() == "yes":
-                  open(file_name, "w").close()
-                  print("All entries deleted.")
-               else:
-                  print("Deletion cancelled.")
+             if confirm.lower() == "yes":
+                open(file_name, "w").close()
+                print("All entries deleted.")
+            else:
+                print("Deletion cancelled.")
 
-        elif choice == 5:
-             print("Goodbye!")
-             break
+     elif choice == 5:
+          print("Goodbye!")
+          break
 
-        else:
-          print("Invalid choice. Please select 1-5.")
+    else:
+        print("Invalid choice. Please select 1-5.")
 
 # Run the program
